+++ CHANGELOG.md
@@ -18,6 +18,8 @@
 - PR #564 Update python `sort_values` method to use updated libcudf `gdf_order_by` API
 - PR #509 CSV Reader: Input CSV file can now be passed in as a text or a binary buffer
 - PR #607 Add `__iter__` and iteritems to DataFrame class
+- PR #576 Create BitMask class and unit test to make 32-bit bit masks
+
 
 ## Improvements